# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agile Project Management API. Manages Projects → User Stories → Tasks for small teams, with asynchronous email notifications.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stories/{storyId}/tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks under a user story
      parameters:
        - $ref: "#/components/parameters/StoryId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task under a user story
      parameters:
        - $ref: "#/components/parameters/StoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{taskId}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task. Triggers an async email notification when assignee or status changes.
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
...
    StoryId:
      name: storyId
      in: path
      required: true
      schema:
        type: integer
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: integer
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Member:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, member]
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    Task:
      type: object
      properties:
        id:
          type: integer
        storyId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, blocked, done]
        assigneeId:
          type: ["integer", "null"]
        assignee:
          oneOf:
            - $ref: "#/components/schemas/Member"
            - type: "null"
        estimateHours:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, storyId, title, description, status, assigneeId, assignee, estimateHours, dueDate, createdAt, updatedAt]
    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        assigneeId:
          type: ["integer", "null"]
        estimateHours:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [todo, in_progress, blocked, done]
          default: todo
      required: [title]
    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [todo, in_progress, blocked, done]
        assigneeId:
          type: ["integer", "null"]
        estimateHours:
          type: ["number", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
...